# Conflicts:
#	live-objects/src/main/kotlin/io/ably/lib/objects/type/BaseLiveObject.kt
#	live-objects/src/main/kotlin/io/ably/lib/objects/type/livecounter/DefaultLiveCounter.kt
#	live-objects/src/main/kotlin/io/ably/lib/objects/type/livecounter/LiveCounterManager.kt
#	live-objects/src/main/kotlin/io/ably/lib/objects/type/livemap/DefaultLiveMap.kt
#	live-objects/src/test/kotlin/io/ably/lib/objects/unit/type/livemap/LiveMapManagerTest.kt

+++ lib/src/main/java/io/ably/lib/objects/type/counter/LiveCounter.java
@@ -58,5 +58,5 @@ public interface LiveCounter extends LiveCounterChange {
      */
     @NotNull
     @Contract(pure = true) // Indicates this method does not modify the state of the object.
-    Long value();
+    Double value();
 }
